+++ heft-plugins/heft-lint-plugin/src/LintPlugin.ts
@@ -15,16 +15,30 @@ import type {
   ITypeScriptPluginAccessor
 } from '@rushstack/heft-typescript-plugin';
 
+import type { LinterBase } from './LinterBase';
 import { Eslint } from './Eslint';
 import { Tslint } from './Tslint';
 import type { IExtendedProgram, IExtendedSourceFile } from './internalTypings/TypeScriptInternals';
 
 const PLUGIN_NAME: 'lint-plugin' = 'lint-plugin';
 const TYPESCRIPT_PLUGIN_NAME: typeof TypeScriptPluginName = 'typescript-plugin';
+const FIX_PARAMETER_NAME: string = '--fix';
 const ESLINTRC_JS_FILENAME: string = '.eslintrc.js';
 const ESLINTRC_CJS_FILENAME: string = '.eslintrc.cjs';
 
-export default class LintPlugin implements IHeftTaskPlugin {
+interface ILintPluginOptions {
+  alwaysFix?: boolean;
+}
+
+interface ILintOptions {
+  taskSession: IHeftTaskSession;
+  heftConfiguration: HeftConfiguration;
+  tsProgram: IExtendedProgram;
+  fix?: boolean;
+  changedFiles?: ReadonlySet<IExtendedSourceFile>;
+}
+
+export default class LintPlugin implements IHeftTaskPlugin<ILintPluginOptions> {
   private readonly _lintingPromises: Promise<void>[] = [];
 
   // These are initliazed by _initAsync
@@ -34,10 +48,25 @@ export default class LintPlugin implements IHeftTaskPlugin {
   private _tslintToolPath: string | undefined;
   private _tslintConfigFilePath: string | undefined;
 
-  public apply(taskSession: IHeftTaskSession, heftConfiguration: HeftConfiguration): void {
+  public apply(
+    taskSession: IHeftTaskSession,
+    heftConfiguration: HeftConfiguration,
+    pluginOptions?: ILintPluginOptions
+  ): void {
     // Disable linting in watch mode. Some lint rules require the context of multiple files, which
     // may not be available in watch mode.
     if (!taskSession.parameters.watch) {
+      let fix: boolean =
+        pluginOptions?.alwaysFix || taskSession.parameters.getFlagParameter(FIX_PARAMETER_NAME).value;
+      if (fix && taskSession.parameters.production) {
+        // Write this as a standard output message since we don't want to throw errors when running in
+        // production mode and "alwaysFix" is specified in the plugin options
+        taskSession.logger.terminal.writeLine(
+          'Fix mode has been disabled since Heft is running in production mode'
+        );
+        fix = false;
+      }
+
       // Use the changed files hook to kick off linting asynchronously
       taskSession.requestAccessToPluginByName(
         '@rushstack/heft-typescript-plugin',
@@ -47,12 +76,13 @@ export default class LintPlugin implements IHeftTaskPlugin {
           accessor.onChangedFilesHook.tap(
             PLUGIN_NAME,
             (changedFilesHookOptions: IChangedFilesHookOptions) => {
-              const lintingPromise: Promise<void> = this._lintAsync(
+              const lintingPromise: Promise<void> = this._lintAsync({
                 taskSession,
                 heftConfiguration,
-                changedFilesHookOptions.program as IExtendedProgram,
-                changedFilesHookOptions.changedFiles as ReadonlySet<IExtendedSourceFile>
-              );
+                fix,
+                tsProgram: changedFilesHookOptions.program as IExtendedProgram,
+                changedFiles: changedFilesHookOptions.changedFiles as ReadonlySet<IExtendedSourceFile>
+              });
               lintingPromise.catch(() => {
                 // Suppress unhandled promise rejection error
               });
@@ -75,7 +105,7 @@ export default class LintPlugin implements IHeftTaskPlugin {
         warningPrinted = true;
 
         // Warn since don't run the linters when in watch mode.
-        taskSession.logger.terminal.writeWarningLine("Linting isn't currently supported in watch mode.");
+        taskSession.logger.terminal.writeWarningLine("Linting isn't currently supported in watch mode");
       } else {
         await Promise.all(this._lintingPromises);
       }
@@ -113,92 +143,53 @@ export default class LintPlugin implements IHeftTaskPlugin {
     }
   }
 
-  private async _lintAsync(
-    taskSession: IHeftTaskSession,
-    heftConfiguration: HeftConfiguration,
-    tsProgram: IExtendedProgram,
-    changedFiles?: ReadonlySet<IExtendedSourceFile>
-  ): Promise<void> {
+  private async _lintAsync(options: ILintOptions): Promise<void> {
+    const { taskSession, heftConfiguration, tsProgram, changedFiles, fix } = options;
+
     // Ensure that we have initialized. This promise is cached, so calling init
     // multiple times will only init once.
     await this._ensureInitializedAsync(taskSession, heftConfiguration);
 
+    const linters: LinterBase<unknown>[] = [];
+    if (this._eslintConfigFilePath && this._eslintToolPath) {
+      const eslintLinter: Eslint = await Eslint.initializeAsync({
+        tsProgram,
+        fix,
+        scopedLogger: taskSession.logger,
+        linterToolPath: this._eslintToolPath,
+        linterConfigFilePath: this._eslintConfigFilePath,
+        buildFolderPath: heftConfiguration.buildFolderPath,
+        buildMetadataFolderPath: taskSession.tempFolderPath
+      });
+      linters.push(eslintLinter);
+    }
+
+    if (this._tslintConfigFilePath && this._tslintToolPath) {
+      const tslintLinter: Tslint = await Tslint.initializeAsync({
+        tsProgram,
+        fix,
+        scopedLogger: taskSession.logger,
+        linterToolPath: this._tslintToolPath,
+        linterConfigFilePath: this._tslintConfigFilePath,
+        buildFolderPath: heftConfiguration.buildFolderPath,
+        buildMetadataFolderPath: taskSession.tempFolderPath
+      });
+      linters.push(tslintLinter);
+    }
+
     // Now that we know we have initialized properly, run the linter(s)
-    const lintingPromises: Promise<void>[] = [];
-    if (this._eslintToolPath) {
-      lintingPromises.push(
-        this._runEslintAsync(
-          taskSession,
-          heftConfiguration,
-          this._eslintToolPath,
-          this._eslintConfigFilePath!,
-          tsProgram,
-          changedFiles
-        )
-      );
-    }
-    if (this._tslintToolPath) {
-      lintingPromises.push(
-        this._runTslintAsync(
-          taskSession,
-          heftConfiguration,
-          this._tslintToolPath,
-          this._tslintConfigFilePath!,
-          tsProgram,
-          changedFiles
-        )
-      );
-    }
-
-    await Promise.all(lintingPromises);
+    await Promise.all(linters.map((linter) => this._runLinterAsync(linter, tsProgram, changedFiles)));
   }
 
-  private async _runEslintAsync(
-    taskSession: IHeftTaskSession,
-    heftConfiguration: HeftConfiguration,
-    eslintToolPath: string,
-    eslintConfigFilePath: string,
+  private async _runLinterAsync(
+    linter: LinterBase<unknown>,
     tsProgram: IExtendedProgram,
     changedFiles?: ReadonlySet<IExtendedSourceFile> | undefined
   ): Promise<void> {
-    const eslint: Eslint = new Eslint({
-      scopedLogger: taskSession.logger,
-      eslintPackagePath: eslintToolPath,
-      linterConfigFilePath: eslintConfigFilePath,
-      buildFolderPath: heftConfiguration.buildFolderPath,
-      buildMetadataFolderPath: taskSession.tempFolderPath
-    });
-
-    eslint.printVersionHeader();
+    linter.printVersionHeader();
 
     const typeScriptFilenames: Set<string> = new Set(tsProgram.getRootFileNames());
-    await eslint.performLintingAsync({
-      tsProgram,
-      typeScriptFilenames,
-      changedFiles: changedFiles || new Set(tsProgram.getSourceFiles())
-    });
-  }
-
-  private async _runTslintAsync(
-    taskSession: IHeftTaskSession,
-    heftConfiguration: HeftConfiguration,
-    tslintToolPath: string,
-    tslintConfigFilePath: string,
-    tsProgram: IExtendedProgram,
-    changedFiles?: ReadonlySet<IExtendedSourceFile> | undefined
-  ): Promise<void> {
-    const tslint: Tslint = new Tslint({
-      scopedLogger: taskSession.logger,
-      tslintPackagePath: tslintToolPath,
-      linterConfigFilePath: tslintConfigFilePath,
-      buildFolderPath: heftConfiguration.buildFolderPath,
-      buildMetadataFolderPath: taskSession.tempFolderPath
-    });
-
-    tslint.printVersionHeader();
-
-    const typeScriptFilenames: Set<string> = new Set(tsProgram.getRootFileNames());
-    await tslint.performLintingAsync({
+    await linter.performLintingAsync({
       tsProgram,
       typeScriptFilenames,
       changedFiles: changedFiles || new Set(tsProgram.getSourceFiles())