+++ libraries/rush-lib/src/cli/RushPnpmCommandLineParser.ts
@@ -31,6 +31,7 @@ import type { IInstallManagerOptions } from '../logic/base/BaseInstallManagerTyp
 import { objectsAreDeepEqual } from '../utilities/objectUtilities';
 import { Utilities } from '../utilities/Utilities';
 import type { Subspace } from '../api/Subspace';
+import type { PnpmOptionsConfiguration } from '../logic/pnpm/PnpmOptionsConfiguration';
 
 const RUSH_SKIP_CHECKS_PARAMETER: string = '--rush-skip-checks';
 
@@ -135,7 +136,7 @@ export class RushPnpmCommandLineParser {
     const subspace: Subspace = rushConfiguration.getSubspace(subspaceName);
     this._subspace = subspace;
 
-    const workspaceFolder: string = subspace.getSubspaceTempFolder();
+    const workspaceFolder: string = subspace.getSubspaceTempFolderPath();
     const workspaceFilePath: string = path.join(workspaceFolder, 'pnpm-workspace.yaml');
 
     if (!FileSystem.exists(workspaceFilePath)) {
@@ -374,7 +375,7 @@ export class RushPnpmCommandLineParser {
 
   private async _executeAsync(): Promise<void> {
     const rushConfiguration: RushConfiguration = this._rushConfiguration;
-    const workspaceFolder: string = this._subspace.getSubspaceTempFolder();
+    const workspaceFolder: string = this._subspace.getSubspaceTempFolderPath();
     const pnpmEnvironmentMap: EnvironmentMap = new EnvironmentMap(process.env);
     pnpmEnvironmentMap.set('NPM_CONFIG_WORKSPACE_DIR', workspaceFolder);
 
@@ -442,7 +443,8 @@ export class RushPnpmCommandLineParser {
       return;
     }
 
-    const subspaceTempFolder: string = this._subspace.getSubspaceTempFolder();
+    const subspaceTempFolder: string = this._subspace.getSubspaceTempFolderPath();
+    const subspaceConfigFolder: string = this._subspace.getSubspaceConfigFolderPath();
 
     switch (commandName) {
       case 'patch-commit': {
@@ -453,7 +455,7 @@ export class RushPnpmCommandLineParser {
         if (this._subspace.getPnpmOptions() === undefined) {
           this._terminal.writeErrorLine(
             `The "rush-pnpm patch-commit" command cannot proceed without a pnpm-config.json file.` +
-              `  Create one in this folder: ${this._subspace.getSubspaceConfigFolder()}`
+              `  Create one in this folder: ${subspaceConfigFolder}`
           );
           break;
         }
@@ -463,13 +465,15 @@ export class RushPnpmCommandLineParser {
         const commonPackageJson: JsonObject = JsonFile.load(commonPackageJsonFilename);
         const newGlobalPatchedDependencies: Record<string, string> | undefined =
           commonPackageJson?.pnpm?.patchedDependencies;
+        const pnpmOptions: PnpmOptionsConfiguration | undefined = this._subspace.getPnpmOptions();
         const currentGlobalPatchedDependencies: Record<string, string> | undefined =
-          this._subspace.getPnpmOptions()?.globalPatchedDependencies;
+          pnpmOptions?.globalPatchedDependencies;
 
         if (!objectsAreDeepEqual(currentGlobalPatchedDependencies, newGlobalPatchedDependencies)) {
           const commonTempPnpmPatchesFolder: string = `${subspaceTempFolder}/${RushConstants.pnpmPatchesFolderName}`;
-          const rushPnpmPatchesFolder: string = `${subspaceTempFolder}/${RushConstants.pnpmPatchesCommonFolderName}`;
-          // Copy (or delete) common\temp\patches\ --> common\pnpm-patches\
+          const rushPnpmPatchesFolder: string = this._subspace.getSubspacePnpmPatchesFolderPath();
+
+          // Copy (or delete) common\temp\subspace\patches\ --> common\config\pnpm-patches\ OR common\config\rush\pnpm-patches\
           if (FileSystem.exists(commonTempPnpmPatchesFolder)) {
             FileSystem.ensureEmptyFolder(rushPnpmPatchesFolder);
             // eslint-disable-next-line no-console
@@ -489,7 +493,7 @@ export class RushPnpmCommandLineParser {
           }
 
           // Update patchedDependencies to pnpm configuration file
-          this._subspace.getPnpmOptions()?.updateGlobalPatchedDependencies(newGlobalPatchedDependencies);
+          pnpmOptions?.updateGlobalPatchedDependencies(newGlobalPatchedDependencies);
 
           // Rerun installation to update
           await this._doRushUpdateAsync();