+++ libraries/rush-lib/src/logic/SetupChecks.ts
@@ -7,7 +7,7 @@ import { FileSystem, AlreadyReportedError } from '@rushstack/node-core-library';
 import { Colorize, PrintUtilities } from '@rushstack/terminal';
 
 import type { RushConfiguration } from '../api/RushConfiguration';
-import { RushConstants } from '../logic/RushConstants';
+import { RushConstants } from './RushConstants';
 
 // Refuses to run at all if the PNPM version is older than this, because there
 // are known bugs or missing features in earlier releases.