+++ libraries/localization-utilities/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log - @rushstack/localization-utilities
 
-This log was last generated on Fri, 02 Aug 2024 17:26:42 GMT and should not be manually modified.
+This log was last generated on Mon, 12 Aug 2024 22:16:04 GMT and should not be manually modified.
+
+## 0.10.0
+Mon, 12 Aug 2024 22:16:04 GMT
+
+### Minor changes
+
+- Update the schema for `.loc.json` files to allow string names that include the `$` character.
 
 ## 0.9.62
 Fri, 02 Aug 2024 17:26:42 GMT