+++ webpack/set-webpack-public-path-plugin/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log - @rushstack/set-webpack-public-path-plugin
 
-This log was last generated on Fri, 02 Aug 2024 17:26:42 GMT and should not be manually modified.
+This log was last generated on Mon, 12 Aug 2024 22:16:04 GMT and should not be manually modified.
+
+## 5.1.47
+Mon, 12 Aug 2024 22:16:04 GMT
+
+_Version update only_
 
 ## 5.1.46
 Fri, 02 Aug 2024 17:26:42 GMT